% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdplot.R
\name{CD_plot}
\alias{CD_plot}
\title{CD plot}
\usage{
CD_plot(X, k)
}
\arguments{
\item{X}{Original data matrix.}

\item{k}{Number of end-members.}
}
\value{
A ggplot object.
}
\description{
Creating coefficient of determination (CD) plot to determine the best number
of end-members (\code{k}).
}
\author{
Shoji F. Nakayama
}
