% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumvar.R
\name{cumvar}
\alias{cumvar}
\title{Cumulative variance plot}
\usage{
cumvar(x)
}
\arguments{
\item{x}{SVD distance}
}
\description{
Plotting cumulative variance plot to determine the best number of end-members (k)
}
\author{
Shoji F. Nakayama
}
