% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumvarplot.R
\name{cumvar_plot}
\alias{cumvar_plot}
\title{Cumulative variance plot}
\usage{
cumvar_plot(x)
}
\arguments{
\item{x}{SVD distance}
}
\description{
Plotting cumulative variance plot to determine the best number of end-members (k)
}
\author{
Shoji F. Nakayama
}
