% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataX}
\alias{dataX}
\title{Dataset X}
\format{
An object of class \code{data.frame} with 119 rows and 9 columns.
}
\usage{
dataX
}
\description{
Dataset X is generated from Matrix A and F
}
\keyword{datasets}
