% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENEG.R
\name{DENEG}
\alias{DENEG}
\title{DENEG}
\usage{
DENEG(x, y, nrows, k)
}
\arguments{
\item{x}{A0}

\item{y}{X estimate}

\item{nrows}{Number of rows in the original matrix X}

\item{k}{Number of end-members}
}
\description{
DENEG
}
\author{
Shoji F. Nakayama
}
