% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xestimate.R
\name{estimate_X}
\alias{estimate_X}
\title{X estimate}
\usage{
estimate_X(X, k)
}
\arguments{
\item{X}{original data matrix}

\item{k}{number of end-members}
}
\description{
Obtain X estimate
}
\author{
Shoji F. Nakayama
}
