% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeF0.R
\name{negative_F0}
\alias{negative_F0}
\title{Negative F0}
\usage{
negative_F0(X, ncols, k)
}
\arguments{
\item{X}{data}

\item{ncols}{Number of columns in the original matrix X}

\item{k}{Number of end-members}
}
\description{
Negative F0
}
\author{
Shoji F. Nakayama
}
