% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgA0}
\alias{orgA0}
\title{Matrix A}
\format{
An object of class \code{data.frame} with 119 rows and 7 columns.
}
\usage{
orgA0
}
\description{
Matrix A is original data from the book to create the dataset X
}
\keyword{datasets}
