% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orgF0}
\alias{orgF0}
\title{Matrix B}
\format{
An object of class \code{data.frame} with 6 rows and 9 columns.
}
\usage{
orgF0
}
\description{
Matrix F is the score matrix from the book to create the sample data X
}
\keyword{datasets}
