% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{PVA}
\alias{PVA}
\title{PVA analysis}
\usage{
PVA(X, k, N = 10, verbose = FALSE)
}
\arguments{
\item{X}{Original data matrix.}

\item{k}{Number of end-members.}

\item{N}{Maximum number of trials.}

\item{verbose}{Logical. If \code{TRUE}, print iteration progress messages.}
}
\value{
A named list with \code{A0} and \code{F0} matrices.
}
\description{
Fit a polytopic vector analysis (PVA) model.
}
\author{
Shoji F. Nakayama
}
