% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultantoblique.R
\name{resultant_oblique}
\alias{resultant_oblique}
\title{Resultant oblique}
\usage{
resultant_oblique(x, y, z, X, ncols, nrows, k)
}
\arguments{
\item{x}{A0}

\item{y}{A11}

\item{z}{O0}

\item{X}{data}

\item{ncols}{Number of columns in the original matrix X}

\item{nrows}{Number of rows in the original matrix X}

\item{k}{Number of end-members}
}
\description{
Resultant oblique
}
\author{
Shoji F. Nakayama
}
