% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varimax.R
\name{varimax}
\alias{varimax}
\title{Varimax}
\usage{
varimax(X, gamma = 1, q = 20, tol = 1e-06)
}
\arguments{
\item{X}{A11}

\item{gamma}{gamma}

\item{q}{q}

\item{tol}{tolerance}
}
\description{
Varimax
}
\author{
Shoji F. Nakayama
}
