% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdplot.R
\name{CD_plot}
\alias{CD_plot}
\title{CD plot}
\usage{
CD_plot(X, k)
}
\arguments{
\item{X}{oritinal data matrix X}

\item{k}{number of end-members}
}
\description{
Creating coeﬀicient of determination (CD) plot to determine the best number of end-members (k)
}
\author{
Shoji F. Nakayama
}
