% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvplot.R
\name{CV_plot}
\alias{CV_plot}
\title{Cumulative variance plot}
\usage{
CV_plot(x)
}
\arguments{
\item{x}{SVD distance}
}
\description{
Plotting cumulative variance plot to determine the best number of end-members (k)
}
\author{
Shoji F. Nakayama
}
