% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DENEG.R
\name{DENEG}
\alias{DENEG}
\title{DENEG}
\usage{
DENEG(X, x, y, k)
}
\arguments{
\item{X}{original data matrix X}

\item{x}{A0}

\item{y}{X_estimate}

\item{k}{number of end-members}
}
\description{
DENEG algorithm
}
\author{
Shoji F. Nakayama
}
