% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatex.R
\name{estimate_X}
\alias{estimate_X}
\title{X estimate}
\usage{
estimate_X(x, y, z, k)
}
\arguments{
\item{x}{SVD data = SVD}

\item{y}{transformed data = X111}

\item{z}{original data = X}

\item{k}{number of end-members}
}
\description{
Obtain X estimate
}
\author{
Shoji F. Nakayama
}
