% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectextreme.R
\name{inspect_extreme}
\alias{inspect_extreme}
\title{Inspect extreme}
\usage{
inspect_extreme(X, Y, Z)
}
\arguments{
\item{X}{A0}

\item{Y}{A11}

\item{Z}{O0}
}
\description{
Inspect extreme
}
\author{
Shoji F. Nakayama
}
