% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectextreme.R
\name{inspect_extreme}
\alias{inspect_extreme}
\title{Inspect extreme}
\usage{
inspect_extreme(A0, A111, O0, k)
}
\arguments{
\item{A0}{A0}

\item{A111}{A111}

\item{O0}{O0}

\item{k}{Number of end-members}
}
\description{
Inspect extreme
}
\author{
Shoji F. Nakayama
}
