% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeA0.R
\name{negative_A0}
\alias{negative_A0}
\title{Negative A0}
\usage{
negative_A0(X, A0, k)
}
\arguments{
\item{X}{original data matrix X}

\item{A0}{A0}

\item{k}{Number of end-members}
}
\description{
Negative A0
}
\author{
Shoji F. Nakayama
}
