% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negativeF0.R
\name{negative_F0}
\alias{negative_F0}
\title{Negative F0}
\usage{
negative_F0(X, Y, k)
}
\arguments{
\item{X}{original data matrix X}

\item{Y}{F0}

\item{k}{number of end-members}
}
\description{
Negative F0
}
\author{
Shoji F. Nakayama
}
