% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{PVA}
\alias{PVA}
\title{PVA analysis}
\usage{
PVA(k, x, y, z, X, N = 10)
}
\arguments{
\item{k}{number of end-members}

\item{x}{X11}

\item{y}{transformed data = X111}

\item{z}{original data = X}

\item{X}{X_estimate}

\item{N}{number of trials}
}
\description{
Polytopic vector analysis (PVA) function
}
\author{
Shoji F. Nakayama
}
