% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva.R
\name{PVA}
\alias{PVA}
\title{PVA analysis}
\usage{
PVA(X, k, N = 10)
}
\arguments{
\item{X}{original data matrix, X}

\item{k}{number of end-members}

\item{N}{number of trials}
}
\description{
Polytopic vector analysis (PVA) function
}
\author{
Shoji F. Nakayama
}
