% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pva4jecs.R
\docType{package}
\name{pva4jecs}
\alias{pva4jecs}
\alias{pva4jecs-package}
\title{pva4jecs}
\description{
This package is a translation from the Python code for polytopic vector analysis (PVA) developed by Dr Yang Ju.
}
