% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultantoblique.R
\name{resultant_oblique}
\alias{resultant_oblique}
\title{Resultant oblique}
\usage{
resultant_oblique(x, y, z, Z, X, k)
}
\arguments{
\item{x}{A111}

\item{y}{F11}

\item{z}{O0}

\item{Z}{X111}

\item{X}{original data matrix X}

\item{k}{number of end-members}
}
\description{
Resultant oblique
}
\author{
Shoji F. Nakayama
}
