% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultantoblique.R
\name{resultant_oblique}
\alias{resultant_oblique}
\title{Resultant oblique}
\usage{
resultant_oblique(X, A0, k)
}
\arguments{
\item{X}{original data matrix X}

\item{A0}{A0}

\item{k}{Number of end-members}
}
\description{
Resultant oblique
}
\author{
Shoji F. Nakayama
}
