% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultantoblique.R
\name{resultant_oblique}
\alias{resultant_oblique}
\title{Resultant oblique}
\usage{
resultant_oblique(A111, F11, O0, X111, k)
}
\arguments{
\item{A111}{A111}

\item{F11}{F11}

\item{O0}{O0}

\item{X111}{X111}

\item{k}{Number of end-members}
}
\description{
Resultant oblique
}
\author{
Shoji F. Nakayama
}
