% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultantoblique.R
\name{resultant_oblique}
\alias{resultant_oblique}
\title{Resultant oblique}
\usage{
resultant_oblique(x, y, z, X)
}
\arguments{
\item{x}{A0}

\item{y}{A11}

\item{z}{O0}

\item{X}{data}
}
\description{
Resultant oblique
}
\author{
Shoji F. Nakayama
}
