% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleback.R
\name{scale_back}
\alias{scale_back}
\title{Transform A11, F11 to original data}
\usage{
scale_back(X, Y, Z, ncols, k, nrows)
}
\arguments{
\item{X}{A11}

\item{Y}{F11}

\item{Z}{X111}

\item{ncols}{Number of columns in the original matrix X}

\item{k}{Number of end-members}

\item{nrows}{Number of rows in the original matrix X}
}
\description{
Transform A11, F11 to original data
}
\author{
Shoji F. Nakayama
}
