% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleback.R
\name{scale_back}
\alias{scale_back}
\title{Transform A11, F11 to original data}
\usage{
scale_back(x, X, Y, Z, k)
}
\arguments{
\item{x}{original data matrix X}

\item{X}{A11}

\item{Y}{F11}

\item{Z}{X111}

\item{k}{Number of end-members}
}
\description{
Transform A11, F11 to original data
}
\author{
Shoji F. Nakayama
}
